# swagger.yaml
swagger: "2.0"
info:
  description: "Сайт школьной системы (ну почти), в котором есть базовый функционал + возсожно регистрация, чего на настоящем сайте школьной системы нельзя, сделано это для возможности использования сайта"
  version: "alpha 1"
  title: "siteCV"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "salvoronis@gmail.com"
  license:
    name: "тут будет лицензия"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /auth/register:
    post:
      tags:
      - "auth"
      summary: "регистрация нового студента (учителей и тд нельзя регистрировать)"
      description: "регистрация"
      operationId: "register"
      consumes: ["application/json"]
      produces: ["application/json"]
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "запрос обработан"
          schema:
            $ref: "#/definitions/Token"
  /school/classes:
    get:
      tags:
      - "school"
      summary: "получить список классов"
      operationId: "getClasses"
      produces:
      - "application/json"
      responses:
        "200":
          description: "запрос обработан"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Classes"
      security:
      - school_auth:
        - "write:pets"
...
securityDefinitions:
  school_auth:
    type: "oauth2"
    authorizationUrl: "localhost:8080/register"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
definitions:
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        example: 200
      type:
        type: "string"
        example: "OK"
      message:
        type: "string"
        example: "All OK. keep working."
...
  Token:
    type: "object"
    properties:
      token:
        type: "string"
      status:
        $ref: "#/definitions/ApiResponse"
  User:
    type: "object"
    properties:
      nickname:
        type: "string"
        example: "salvoroni"
      password:
        type: "string"
        example: "qwerty"
      class:
        type: "integer"
        example: 1
        description: "class id > 0"
      firstName:
        type: "string"
        example: "Danil"
      secondName:
        type: "string"
        example: "Dickens"
      termOfUse:
        type: "boolean"
        default: false
      email:
        type: "string"
        example: "danilDick@mail.ru"
  Classes:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "11 A"
...